apps/dashboard/api-service/pyproject.toml
[tool.poetry]
name = "codegen"
version = "0.1.0"
description = ""
authors = ["e2b <hello@e2b.dev>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dateutil = "^2.8.2"
urllib3 = "^1.26.14"
pydantic = "^1.10.6"
aenum = "^3.1.11"
asyncio = "^3.4.3"
postgrest = "^0.10.6"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
fastapi = "^0.95.1"
langchain = "0.0.141"
openai = "^0.27.6"
uuid = "^1.30"
aiohttp = "^3.8.4"
gunicorn = "^20.1.0"
opentelemetry-distro = "^0.39b0"
opentelemetry-exporter-otlp = "^1.18.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"
